+++ Wrappers/Python/test/test_algorithms.py
@@ -32,7 +32,7 @@ from cil.framework.labels import FillType
 
 from cil.optimisation.utilities import ArmijoStepSizeRule, ConstantStepSize, Sampler, callbacks, Sensitivity, StepSizeRule
 from cil.optimisation.algorithms.APGD import NesterovMomentum, ScalarMomentumCoefficient, ConstantMomentum
-from cil.optimisation.operators import IdentityOperator
+from cil.optimisation.operators import IdentityOperator, AdjointOperator
 from cil.optimisation.operators import GradientOperator, BlockOperator, MatrixOperator
 
 
@@ -861,6 +861,72 @@ class TestCGLS(CCPiTestClass):
 
 class TestPDHG(CCPiTestClass):
 
+    def setUp(self):
+        self.ig = ImageGeometry(10, 10)
+        self.initial = self.ig.allocate(0)
+        self.data = self.ig.allocate('random')
+        self.operator = IdentityOperator(self.ig)
+
+        self.f = L2NormSquared(b=self.data)
+        self.g = ZeroFunction()
+
+        self.subsets = 10
+
+        data = dataexample.SIMULATED_PARALLEL_BEAM_DATA.get(size=(16, 16))
+
+        self.partitioned_data = data.partition(self.subsets, 'sequential')
+        self.A_block = BlockOperator(
+            *[IdentityOperator(self.partitioned_data[i].geometry) for i in range(self.subsets)])
+
+        # block function
+        self.F_block = BlockFunction(*[L2NormSquared(b=self.partitioned_data[i])
+                                for i in range(self.subsets)])
+        alpha = 0.025
+        self.G = alpha * IndicatorBox(lower=0)
+        
+        self.A_block_adjoint = AdjointOperator(self.A_block)
+        self.f_adjoint = L2NormSquared(b=self.data)
+        
+    def test_init_primal_dual(self):
+        init1 = self.ig.allocate(1)
+        init2 = self.ig.allocate(2)
+        init_default = self.ig.allocate(0)
+        pdhg = PDHG(initial=None, f=self.f, g=self.g,
+                         operator=self.operator)
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.as_array(), init_default.as_array())
+        self.assertNumpyArrayAlmostEqual(pdhg.y.as_array(), init_default.as_array())
+        
+        pdhg = PDHG(initial=init1, f=self.f, g=self.g,
+                         operator=self.operator)
+        
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.as_array(), init1.as_array())
+        self.assertNumpyArrayAlmostEqual(pdhg.y.as_array(), init_default.as_array())
+        
+        pdhg = PDHG(initial=(None, init2), f=self.f, g=self.g,
+                         operator=self.operator) # check tuple initialisation just dual
+        
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.as_array(), init_default.as_array())
+        self.assertNumpyArrayAlmostEqual(pdhg.y.as_array(), init2.as_array())
+        
+        pdhg= PDHG(initial=self.partitioned_data, f=self.f_adjoint, g=self.G,
+                         operator=self.A_block_adjoint) # check can initialise with block data without dual initialisation
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.get_item(0).as_array(), self.partitioned_data.get_item(0).as_array())
+        self.assertEqual(np.sum(pdhg.y.as_array()), 0)
+        
+        pdhg = PDHG(initial=[init1, None], f=self.f, g=self.g,
+                         operator=self.operator)  # check list initialisation just primal
+        
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.as_array(), init1.as_array())
+        self.assertNumpyArrayAlmostEqual(pdhg.y.as_array(), init_default.as_array())
+        
+        pdhg = PDHG(initial=[init1, init2], f=self.f, g=self.g,
+                         operator=self.operator)  # check list initialisation both primal and dual
+        
+        self.assertNumpyArrayAlmostEqual(pdhg.x_old.as_array(), init1.as_array())
+        self.assertNumpyArrayAlmostEqual(pdhg.y.as_array(), init2.as_array())
+        
+        
+        
     def test_PDHG_Denoising(self):
         # adapted from demo PDHG_TV_Color_Denoising.py in CIL-Demos repository
         data = dataexample.PEPPERS.get(size=(256, 256))